requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.9
djangorestframework==3.12.4
flake8==4.0.1
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
